refactor: file layouts
